refactor: extract capability.lua out of engine.lua
Terminal-support detection (async/sync probe + state) moves out of the
placement engine into svgtree.capability, so the engine is just placement and
support detection lives in one module. health.lua stops probing the terminal
itself: it calls the new terminal-only capability.terminal_supported() for its
terminal line and keeps its separate converter line, so its output is unchanged.
The APC reply parser is extracted as a pure parse_apc_reply() and covered by a
headless smoke script (scripts/test.sh + scripts/test-capability.lua).
